LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/responsive-iframe/responsive-iframe-navigation-unsized_origin=cross-expected.txt



FAIL Responsive iframe clears size during navigation to non-responsive page assert_equals: Initial height should be 200px expected 200 but got 150
